Try to prevent identical version numbers across branches by matching hardcoded prerelease identifiers with hardcoded branch name.
If we hardcode a prelease identifier for a .NET release, we can easily end up
with multiple branches using the same version number, because the hardcoded
prerelease identifier ends up being the same for any other branches that are
created later from the release branch (one typical example would be backport
branches).

This is bad.

So try to prevent this, by hardcoding two values instead!

We'll hardcode the branch name where the hardcoded prerelease identifier
should be applied, and this way any other branches will not use the hardcoded
prerelease identifier.

Ref: https://github.com/xamarin/xamarin-macios/pull/11891.