[static registrar] Add support for manually passing assemblies to be registered to the static registrar. Fixes #11309.
Add support for manually adding assemblies to the list of assemblies processed by
the static registrar.

The assemblies are specified by adding an MSBuild item to the project:

    <StaticRegistrarReference Include="path/to/assembly.dll" />

The item can also specify a location in the app bundle where the asssembly should
be placed by the build by using the Link metadata:

    <StaticRegistrarReference Include="path/to/assembly.dll" Link="relative/path" />

'Link' is a path relative to where the other assemblies are located in the app (the
root directory on mobile platforms, and in Contents/MonoBundle for macOS and Mac
Catalyst).

This is implemented both for legacy Xamarin and .NET.

Fixes https://github.com/xamarin/xamarin-macios/issues/11309.