[d16-10][iad] Remove most of the framework
Most of the API were removed in Xcode 13 beta 1.

The App Store also started to reject some of the (already deprecated)
selectors that were removed, making the API not usable anymore (even
on older versions of the OS).

This is a manual backport of https://github.com/xamarin/xamarin-macios/pull/11906
as the header for Xcode 12.5 are different (and xtro would fail based
on them)