[AppKit] Make the files containing enums API source files instead of Core source files.
This makes it easier to include/exclude certain enums in Mac Catalyst (since
we can use [NoMacCatalyst] instead of littering the files with "#if
MACCATALYST").

It required moving a little bit of unrelated code out of the enums source
files, but this is a much smaller change than moving all of the enums out of
these files (see #12185 for how that ends up).