[msbuild] Move the CreateIpa target to shared code.
There was a main version of this target for Xamarin.iOS, and then numerous dummy targets for two scenarios:

* Building app extensions.
* Building macOS apps.

The first one is handled by the '_CanArchive' property (will be false when
building app extensions), and for the second case I've added it to the
condition on the CreateIpa target.

This way we don't have to have logic elsewhere about when it's valid to call
CreateIpa.