[dotnet-tools] Fix Profile.Is[Product|Sdk]Assembly. Fixes #12276.
In mtouch and mmp's implementation of these functions, they take the assembly
name, not the assembly path.

So for .NET, we'd previouslyget input such as "Xamarin.iOS", think it was a
filename, strip the extension, and compare "Xamarin" to the platform assembly
name, which didn't work.

Fixes https://github.com/xamarin/xamarin-macios/issues/12276.