[xharness] Automatically inline Imports to a "shared.csproj". Fixes #12316 and #12277.
This way we process the shared.csproj that we use in our .NET test projects. The
main effect is that we'll clone any project references included in the shared.csproj,
which on one hand will fix #12316, and on another it means that we'll add/remove
defines in referenced projects, which fixes #12277.

Fixes https://github.com/xamarin/xamarin-macios/issues/12316.
Fixes https://github.com/xamarin/xamarin-macios/issues/12277.