[tools] Link away Runtime.RegisterEntryAssembly if and only if we've optimized away the dynamic registrar. Fixes #12327.
Previously we'd only call Runtime.RegisterEntryAssembly in the simulator if
the dynamic registrar was available, but now we may call it on device as well
(still only if the dynamic registrar is available). So modify the linker to
keep Runtime.RegisterEntryAssembly even if we're executing on device, as long
as the dynamic registrar is around.

This ensures we get the same behavior both in the simulator and on device (and
desktop for that matter).

Fixes https://github.com/xamarin/xamarin-macios/issues/12327.