[src] Define __MACOS__ for macOS builds.
This also meant I had to remove Obsolete attributes when there's an existing
Obsoleted attribute.

For .NET we'll generate an Obsolete attribute from any Obsoleted attribute,
and if there already are any Obsolete attributes as well, then this happens:

    build/dotnet/macos/generated-sources/Intents/INCallCapability.g.cs(49,3): error CS0579: Duplicate 'Obsolete' attribute
    build/dotnet/macos/generated-sources/Intents/INCallCapabilityOptions.g.cs(50,3): error CS0579: Duplicate 'Obsolete' attribute
    build/dotnet/macos/generated-sources/Intents/INCallDestinationType.g.cs(50,3): error CS0579: Duplicate 'Obsolete' attribute
    build/dotnet/macos/generated-sources/Intents/INCallRecordTypeOptions.g.cs(50,3): error CS0579: Duplicate 'Obsolete' attribute
    [...]

So avoid this by removing any Obsolete attributes when there's also an
Obsoleted attribute.