[Makefile] Prevent people from bulding in the wrong dir.
Special paths are protected by the OS. This paths are problematic
becasue the xpcproxy_sim might not have access to them, resulting in the
simulator tests failing (the process cannot return the PID used to
launch the app).

The problem with special folders.. is that they could be or not
translated or changed so we cannot hardcode them, yet we can use a small
apple script to return the path of the ones we know about (documents,
downloads, etc), get them and check if pwd is a substring.