[msbuild] Fix computation of min OS version for Mac Catalyst.
We store the macOS version in the Info.plist, and use the iOS version for the MinimumOSVersion
MSBuild variable.

This means that in ReadAppManifest we must convert the min OS version from the Info.plist
from a macOS version to an iOS version (and not do it in CompileAppManifest).

Also add support for the iOS version in the Info.plist for Mac Catalyst, and automatically
convert it to the corresponding macOS version.