[dotnet] Support SupportedOSPlatformVersion. Fixes #12336.
* Add support for the SupportedOSPlatformVersion MSBuild property, and write it
  to the Info.plist for the corresponding minimum OS version.
* If there are any minimum OS version in the Info.plist, and show an error if it
  doesn't match SupportedOSPlatformVersion.

This unfortunately means that if there's any minimum OS version in any Info.plist,
then that will most likely have to be moved to the SupportedOSPlatformVersion property
(or removed entirely if that's the right choice), since it's unlikely to match the
default value for SupportedOSPlatformVersion. However, this was deemed to be the
best option for the future (it's a one-time pain during migration).

Fixes https://github.com/xamarin/xamarin-macios/issues/12336.