[dotnet] Don't use the built-in publishing logic in .NET to copy frameworks to the app bundle. Fixes #12369.
.NET/MSBuild don't handle symlinks properly [1], which means that we can't ask .NET
to copy frameworks to the app bundle, since frameworks may contain symlinks.

In our case, the symptom was that instead of copying symlinks, the file the symlink
pointed to was copied instead, and then codesign complained about invalid bundle
format when we tried to sign the framework.

We fix this by having our own target (_CopyFrameworksToBundle) to copy frameworks
to the app bundle (instead of adding all the files in the frameworks to the ResolvedFileToPublish
item group), and then using 'ditto' to copy the frameworks.

[1]: https://github.com/dotnet/msbuild/issues/6821

Fixes https://github.com/xamarin/xamarin-macios/issues/12369.