[msbuild] Fix signing frameworks where the executable is a symlink.
We were first resolving symlinks for the input - say we had an Example.framework/Example
symlink to Example.framework/Versions/A/Example - and then checking the parent directory
if it's a framework. The parent directory of 'Example.framework/Versions/A/Example'
is 'A', which did not meet our framewrok condition (if it ends with '.framework').

Adjust the logic to resolve symlinks after checking if the input is a framework or
not.