[msbuild] Add support for the ScnTool task on Mac Catalyst.
It seems the argument Xcode passes to scntool has changed since our initial implementation,
and since updating to the new arguments makes things for Mac Catalyst, let's do just
that.

I don't know which version of Xcode changed the arguments to scntool, but it's no
later than the current Xcode (13), so I've used that as the version check.