[msbuild/dotnet] Compile project-level assets in the outer build for multi-rid builds. Fixes #12410.
We don't need to compile project-level assets for every RuntimeIdentifier in multi-rid
builds, we can instead compile them just once in the outer build.

There is also a correctness issue here: we can't compile assets more than once and
expect to get the exact same compiled result every time (in particular actool seems
to be adding random bytes in to the compiled output), and this creates a problem
when trying to merge the different runtime-specific compiled output into a universal
binary.

We accomplish this by:

* Processing these assets in the outer build, before we execute the rid-specific
  inner builds.
* Store the paths to the assets we've processed in a file.
* In the inner builds, we read that file, and remove any matches from the corresponding
  item group.
* Make sure to copy the compiled assets to the app bundle at the end of the outer
  build.

These are the assets we currently handle this way:

* ImageAsset
* InterfaceDefinition
* SceneKitAsset
* Collada
* TextureAtlas
* CoreMLModel

Fixes https://github.com/xamarin/xamarin-macios/issues/12410.