[mlaunch] Make sure to install mlaunch when the 'install' target is executed. (#12801)
When we create a package, we only execute 'make install' with a custom install
directory (and not 'make all'). This means that we have to make sure 'make
install' actually copies mlaunch to the install directory, otherwise it won't
end up in the package.

So make sure to do this when we're using mlaunch from macios-binaries (it's
already handled correctly in maccore/tools/mlaunch when building mlaunch from
source).